Share Markdown rendering between CLI and TUI
Field-scoped CLI rendering needs the TUI's established terminal style and ANSI-safe wrapping behavior without maintaining a second renderer. Centralizing that boundary lets both surfaces sanitize stored content and preserve readable wrapping while the TUI continues to supply its adaptive code-block background.\n\nGenerated with Codex\n\nCo-authored-by: Codex <codex@openai.com>
